+++ .gitignore
@@ -1,5 +1,9 @@
 # See https://help.github.com/articles/ignoring-files/ for more about ignoring files.
 
+
+#environment
+./env
+
 # dependencies
 /node_modules
 /.pnp
