+++ src/App.js
@@ -1,36 +1,36 @@
 import React from "react";
 import { Route, Routes } from "react-router-dom";
 import "./App.css";
-import Home from "./Navigation/Home";
-import NewsInfo from "./Navigation/NewsInfo";
-import News from "./Navigation/News";
-import CardInfo from "./Navigation/CardInfo";
-import Portfolio from "./Navigation/Portfolio";
-import WatchLIst from "./Navigation/WatchLIst";
-import Search from "./Navigation/Search";
-import Exchange from "./Navigation/Exchange";
-import MainHome from "./Navigation/MainHome";
-import Settings from "./Navigation/Settings";
-import { ContextProvider } from "./Store/Context/AuthContext";
-import ErrorPage from "./Navigation/ErrorPage";
-import UserProfile from "./Navigation/UserProfile";
-import LoginPage from "./Navigation/LoginPage";
-import { PrivateRoutes } from "./Store/ProtectedRoute/ProtectedRoute";
+import Home from "./Pages/index.jsx";
+import HomeLayout from "./Layout/HomeLayout.jsx";
+import News from "./Pages/News/News.jsx";
+import NewsInfo from "./Pages/News/NewsInfo.jsx";
+import CoinInfo from "./Pages/CoinInfo.jsx";
+import Portfolio from "./Pages/Portfolio";
+import WatchLIst from "./Pages/WatchLIst.jsx";
+import Search from "./Pages/Search";
+import Exchange from "./Pages/Exchange";
+import Settings from "./Pages/Settings";
+import { ContextProvider } from "./Store/AuthContext";
+import ErrorPage from "./Pages/Error/404.jsx";
+import UserProfile from "./Pages/UserProfile";
+import LoginPage from "./Pages/LoginPage";
+import { PrivateRoutes } from "./Layout/ProtectedRoute";
 
 function App() {
   return (
     <ContextProvider>
       <Routes>
         <Route element={<PrivateRoutes />}>
-          <Route path="/" element={<Home />}>
-            <Route index element={<MainHome />} />
+          <Route path="/" element={<HomeLayout />}>
+            <Route index element={<Home />} />
             <Route path="/watchlist" element={<WatchLIst />} />
             <Route path="/news" element={<News />} />
             <Route path="/exchange" element={<Exchange />} />
           </Route>
           <Route path="/search" element={<Search />} />
           <Route path="/Portfolio" element={<Portfolio />} />
-          <Route path="/cardinfo/:id" element={<CardInfo />} />
+          <Route path="/CoinInfo/:id" element={<CoinInfo />} />
           <Route path="/newsinfo/:id" element={<NewsInfo />} />
           <Route path="/settings" element={<Settings />} />
           <Route path="/profile" element={<UserProfile />} />
